Account Magic
Converted ProductPage to PHP.
Can now add products to the Cutomer's favourites; also shows if a
product is favourited on the product page.
Header now changes Login/Register to My Account when logged in.
*NOTE* All pages must now start with session_start() for this to work
properly. *END NOTE*
Login and CreateAccount pages now redirect if the user is already logged
in.